chore(zql): allow source tests to be run against any source implementationt
